requirements.txt
langchain==0.1.9
langchain-community==0.0.24
torch==2.2.1
transformers==4.38.2
pillow==10.2.0